refactor: move auth from `core-services/authorization` to `plugins/simple-authorization`
