+++ src/plugins/simple-authorization/index.js
