+++ src/plugins/simple-authorization/util/hasPermission.js
