+++ src/plugins/simple-authorization/util/hasPermission.test.js
