+++ src/plugins/simple-authorization/util/shopsUserHasPermissionFor.js
