+++ src/plugins/simple-authorization/util/shopsUserHasPermissionFor.test.js
